project/requirements.txt
metadrive-simulator
numpy
opencv-python
pandas

torch
torchvision
